requirements.txt
PyYAML>=6.0.0
munch>=4.0.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.0.0
matplotlib>=3.8.0
scipy>=1.11.0
tqdm>=4.66.0
seaborn>=0.13.0
